requirements.txt
# The Adaptive Chef - Requirements
# Python 3.10+ required

# Core ADK and Gemini dependencies
google-genai>=1.0.0
google-adk>=0.1.0

# Additional utilities (if needed)
python-dotenv>=1.0.0
pydantic>=2.0.0

# For deployment (optional)
gunicorn>=21.0.0
uvicorn>=0.24.0
